Revert "Scene, Scene Manager, PlayScene 클래스 분할 했으나 정상 작동되지 않음."
This reverts commit db1a074cef3817d6d1180e30182795cdd7a3679a.

+++ DuhiEngine_Source/DHGameObject.cpp
@@ -68,7 +68,7 @@ namespace dh {
          //Rectangle(mHdc, 100 + mX, 100 + mY, 200 + mX, 200 + mY);
 
         
-        Rectangle(hdc, 100 + (int)mX, 200 + (int)mY, 200 + (int)mX, 300 + (int)mY);
+        Rectangle(hdc, 100 + mX, 200 + mY, 200 + mX, 300 + mY);
 
         SelectObject(hdc, oldBrush);
         /*DeleteObject(grayBrush);*/
